pyproject.toml
[project]
name = "hn-local-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "mflux>=0.17.4",
    "mlx-lm>=0.31.1",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "typer>=0.24.1",
]
